point selection file name change
